+++ cabal-testsuite/PackageTests/Regression/T5409/repo/build-tool-pkg-2/main/Main.hs
@@ -1,0 +1,11 @@
+module Main where
+
+import BuildToolLibrary (buildToolLibraryVersion)
+import System.Environment
+
+main = do
+  (_:source:target:_) <- getArgs
+  writeFile target . unlines . map replaceVersion . lines =<< readFile source
+
+replaceVersion "    BUILD_TOOL_VERSION" = "    " ++ show buildToolLibraryVersion
+replaceVersion line                     = line