+++ cabal-testsuite/PackageTests/Regression/T5409/repo/build-tool-pkg-2/src/BuildToolLibrary.hs
@@ -1,0 +1,4 @@
+module BuildToolLibrary where
+
+buildToolLibraryVersion :: Int
+buildToolLibraryVersion = 2